Migrated auction house handling to TS

+++ helperFuncs/auctionHandler.ts
@@ -1,6 +1,8 @@
 import os from "os";
+import type {AuctionResponse, AuctionProduct} from "../types/priceHandlers.ts";
+import type {AuctionWorkerInput, AuctionWorkerOutput} from "../types/workers.ts";
 
-function chunkInto(array, numChunks = 1) {
+function chunkInto(array: number[], numChunks: number = 1): number[][] {
     const chunkSize = Math.ceil(array.length / numChunks);
     const pageChunks = [];
     for (let i = 0; i < array.length; i += chunkSize) {
@@ -9,19 +11,19 @@ function chunkInto(array, numChunks = 1) {
     return pageChunks;
 }
 
-export async function fetchAuctionPrices(neededItems) {
+export async function fetchAuctionPrices(neededItems: string[]) {
     const auctionUrl = new URL("https://api.hypixel.net/v2/skyblock/auctions");
     const auctionResponse = await fetch(auctionUrl);
-    const auctionPageContent = await auctionResponse.json();
+    const auctionPageContent = await auctionResponse.json() as AuctionResponse;
+    if(!auctionPageContent.success) return new Map();
     const pageAmount = auctionPageContent["totalPages"];
-
     const pageChunks = chunkInto(Array.from(Array(pageAmount), (_, i) => i), os.cpus().length);
     const scatteredPrices = await Promise.all(pageChunks.map(pages => {
         const worker = new Worker(new URL("./auctionWorker.js", import.meta.url));
 
-        return new Promise((resolve, reject) => {
-            worker.postMessage({pages, neededItems});
-            worker.onmessage = e => { resolve(e.data); worker.terminate(); };
+        return new Promise<AuctionWorkerOutput>((resolve, reject) => {
+            worker.postMessage({pages, neededItems} satisfies AuctionWorkerInput);
+            worker.onmessage = (event: MessageEvent<AuctionWorkerOutput>) => { resolve(event.data); worker.terminate(); };
             worker.onerror = e => { reject(e); worker.terminate(); };
         });
     }));
@@ -33,7 +35,7 @@ export async function fetchAuctionPrices(neededItems) {
             existingValue ? existingValue.push(...value) : acc.set(modifiedKey, value);
         }
         return acc;
-    }, new Map());
+    }, new Map<string, number[]>());
 
     for (const [key, value] of auctionItemsPrices) {
         auctionItemsPrices.set(key, value.sort((a, b) => a - b));
